moved models into src, recreated User Scheam
