+++ .sequelizerc
@@ -1,8 +1,8 @@
 var path = require('path');
 
 module.exports = {
-  'config': path.resolve('./server', 'config.json'),
-  'migrations-path': path.resolve('./server', 'migrations'),
-  'models-path': path.resolve('./server', 'models'),
-  'seeders-path': path.resolve('./server', 'seeders')
+  'config': path.resolve('./src/server', 'config.json'),
+  'migrations-path': path.resolve('./src/server', 'migrations'),
+  'models-path': path.resolve('./src/server', 'models'),
+  'seeders-path': path.resolve('./src/server', 'seeders')
 }
