+++ src/server/config.json
@@ -12,8 +12,12 @@
     "database": "guest_app_test",
     "host": "127.0.0.1",
     "dialect": "postgres"
-  },  
+  },
   "production": {
-    "use_env_variable": "DATABASE_URL"
+    "username": "root",
+    "password": null,
+    "database": "database_production",
+    "host": "127.0.0.1",
+    "dialect": "mysql"
   }
 }
